split logging into its own file
